deprecate Menu and rename it to DropdownMenu
