+++ src/__screenshot_tests__/__image_snapshots__/dropdown-menu-screenshot-test-tsx-dropdown-menu-positioned-at-top-left-2-snap.png
