+++ src/__screenshot_tests__/__image_snapshots__/dropdown-menu-screenshot-test-tsx-dropdown-menu-respect-max-height-when-options-overflow-and-positioned-at-top-right-1-snap.png
